+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - Privatized renderer backend
 - Small optimizations on renderer sorting and copying sprites
 - Expanded Texture loading to instead use `png` and `zune_jpeg`
+- Updated crates to latest versions
 
 ## v0.1.0 (24/07/2026)